// EmployeeManager.ClientApp/src/app/core/api/generated/model/countryUpdateDto.ts
/**
 * EmployeeManager.WebAPI
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface CountryUpdateDto { 
    name?: string | null;
    code?: string | null;
}