// EmployeeManager.ClientApp/src/app/core/api/generated/model/cityUpdateDto.ts
/**
 * EmployeeManager.WebAPI
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface CityUpdateDto { 
    name?: string | null;
    countryId?: number | null;
}