// EmployeeManager.ClientApp/src/app/core/api/generated/model/addressCreateDto.ts
/**
 * EmployeeManager.WebAPI
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface AddressCreateDto { 
    street?: string | null;
    houseNumber?: string | null;
    zipCode?: string | null;
    cityId?: number | null;
    countryId?: number | null;
}